Relocate branchSwitcher componet to own file
