+++ packages/airview-cms/src/features/editor/top-bar.js
@@ -1,16 +1,7 @@
-import React, { useEffect } from "react";
-import { useSelector, useDispatch } from "react-redux";
-import {
-  AppBar,
-  Toolbar,
-  Button,
-  Box,
-  Typography,
-  Select,
-  MenuItem,
-} from "@mui/material";
-import { setWorkingBranch, fetchBranches } from "./branch-slice";
+import React from "react";
+import { AppBar, Toolbar, Button, Box, Typography } from "@mui/material";
 import { version } from "../../../package.json";
+import { BranchSwitcher } from "./branch-switcher";
 
 export function TopBar() {
   return (
@@ -64,6 +55,7 @@ export function TopBar() {
                   lineHeight: 1.2,
                   textTransform: "uppercase",
                   fontSize: 11,
+                  fontWeight: "medium",
                 }}
               >
                 Version: {version}
@@ -99,7 +91,7 @@ export function TopBar() {
       >
         <Box
           sx={{
-            flex: "1 1 auto",
+            flex: "0 1 auto",
           }}
         >
           <BranchSwitcher />
@@ -114,14 +106,9 @@ export function TopBar() {
             },
           }}
         >
-          {/*<Button
-            variant="text"
-            size="small"
-            disabled={Boolean(uiSlice.activeModalId)}
-            onClick={() => dispatch(setActiveModalId(BRANCH_SWITCHER_MODAL_ID))}
-          >
-            Switch Branch
-          </Button>*/}
+          <Button variant="text" size="small">
+            Edit Meta
+          </Button>
           <Button variant="text" size="small">
             Create Branch
           </Button>
@@ -137,56 +124,5 @@ export function TopBar() {
         </Box>
       </Toolbar>
     </AppBar>
-  );
-}
-
-function BranchSwitcher() {
-  const dispatch = useDispatch();
-  const branchesSlice = useSelector((state) => state.branchesSlice);
-  const { status, workingBranch, branches } = branchesSlice;
-
-  useEffect(() => {
-    dispatch(fetchBranches());
-  }, [dispatch]);
-
-  const getValue = () => {
-    if (status === "loading" || status === "idle" || status === "error") {
-      return "placeholder";
-    } else {
-      return workingBranch;
-    }
-  };
-
-  return (
-    <Select
-      labelId="demo-simple-select-label"
-      id="demo-simple-select"
-      value={getValue()}
-      sx={{
-        height: 25,
-        "& .MuiSelect-select": {
-          fontSize: 13,
-          paddingTop: 1,
-          paddingBottom: 1,
-        },
-      }}
-      disabled={status === "loading" || status === "idle" || status === "error"}
-      onChange={(event) => dispatch(setWorkingBranch(event.target.value))}
-    >
-      {status !== "fulfilled" && (
-        <MenuItem value="placeholder" dense>
-          {status === "error"
-            ? "Error loading branches!"
-            : "Loading branches..."}
-        </MenuItem>
-      )}
-
-      {branches &&
-        branches.map((branch) => (
-          <MenuItem value={branch.name} key={branch.name} dense>
-            {branch.name}
-          </MenuItem>
-        ))}
-    </Select>
   );
 }
